requirements.txt
nextcord
requests
aiohttp
python-dotenv
sqlite3